Add -Wconversion and -Wsign-conversion and fix resulting issues

+++ andino_firmware/include/andino/app/hw.h
@@ -29,42 +29,44 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
+#include <stdint.h>
+
 namespace andino {
 
 /// @brief Hardware configuration.
 struct Hw {
   /// @brief Left encoder channel A pin. Connected to PD2 (digital pin 2).
-  static constexpr int kLeftEncoderChannelAGpioPin{2};
+  static constexpr uint8_t kLeftEncoderChannelAGpioPin{2};
   /// @brief Left encoder channel B pin. Connected to PD3 (digital pin 3).
-  static constexpr int kLeftEncoderChannelBGpioPin{3};
+  static constexpr uint8_t kLeftEncoderChannelBGpioPin{3};
 
   /// @brief Right encoder channel A pin. Connected to PC2 (digital pin 16, analog pin A2).
-  static constexpr int kRightEncoderChannelAGpioPin{16};
+  static constexpr uint8_t kRightEncoderChannelAGpioPin{16};
   /// @brief Right encoder channel B pin. Connected to PC3 (digital pin 17, analog pin A3).
-  static constexpr int kRightEncoderChannelBGpioPin{17};
+  static constexpr uint8_t kRightEncoderChannelBGpioPin{17};
 
   /// @brief Left motor driver backward pin. Connected to PD6 (digital pin 6).
-  static constexpr int kLeftMotorBackwardGpioPin{6};
+  static constexpr uint8_t kLeftMotorBackwardGpioPin{6};
   /// @brief Left motor driver forward pin. Connected to PB2 (digital pin 10).
-  static constexpr int kLeftMotorForwardGpioPin{10};
+  static constexpr uint8_t kLeftMotorForwardGpioPin{10};
   /// @brief Left motor driver enable pin. Connected to PB5 (digital pin 13).
   /// @note The enable input of the L298N motor driver may be directly jumped to 5V if the board has
   /// a jumper to do so.
-  static constexpr int kLeftMotorEnableGpioPin{13};
+  static constexpr uint8_t kLeftMotorEnableGpioPin{13};
 
   /// @brief Right motor driver backward pin. Connected to PD5 (digital pin 5).
-  static constexpr int kRightMotorBackwardGpioPin{5};
+  static constexpr uint8_t kRightMotorBackwardGpioPin{5};
   /// @brief Right motor driver forward pin. Connected to PB1 (digital pin 9).
-  static constexpr int kRightMotorForwardGpioPin{9};
+  static constexpr uint8_t kRightMotorForwardGpioPin{9};
   /// @brief Right motor driver enable pin. Connected to PB4 (digital pin 12).
   /// @note The enable input of the L298N motor driver may be directly jumped to 5V if the board has
   /// a jumper to do so.
-  static constexpr int kRightMotorEnableGpioPin{12};
+  static constexpr uint8_t kRightMotorEnableGpioPin{12};
 
   /// @brief IMU sensor I2C SCL pin. Connected to PC5 (digital pin 19, analog pin A5).
-  static constexpr int kImuI2cSclPin{19};
+  static constexpr uint8_t kImuI2cSclPin{19};
   /// @brief IMU sensor I2C SDA pin. Connected to PC4 (digital pin 18, analog pin A4).
-  static constexpr int kImuI2cSdaPin{18};
+  static constexpr uint8_t kImuI2cSdaPin{18};
 };
 
 }  // namespace andino
